# pkg/server/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Marionette API
  version: 1.0.0
  description: |-
    The public Marionette API: create sessions, run tasks on them, answer the
    agent's permission requests, and stream the output.

    Authenticate with an API key as `Authorization: Bearer mk_...`. Each route
    requires a scope, listed in its security requirement. The two WebSocket routes
    also accept the key as a `?token=` query parameter, because a browser cannot
    set headers on a WebSocket handshake.

    List endpoints are cursor-paginated: pass `next_cursor` from one response as
    `cursor` on the next request, and stop when `has_more` is false. Repeated
    filters such as `status` are sent as repeated keys (`?status=a&status=b`),
    not as `status[]`.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/sessions/{sessionID}/tunnels:
    post:
      tags:
        - Tunnels
      summary: Open a tunnel into a session
      description: Forwards a port inside the runner through the API server. The response is the only place the tunnel token is readable.
      operationId: postSessionsTunnels
      security:
        - ApiKeyAuth:
            - tunnels:write
      parameters:
        - name: sessionID
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTunnelRequest'
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tunnel'
        "401":
          description: The API key is missing, malformed, revoked or expired.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: The API key lacks the tunnels:write scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: No such resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        default:
          description: An error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
        - Tunnels
      summary: List a session's tunnels
      operationId: getSessionsTunnels
      security:
        - ApiKeyAuth:
            - tunnels:read
      parameters:
        - name: sessionID
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TunnelList'
        "401":
          description: The API key is missing, malformed, revoked or expired.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: The API key lacks the tunnels:read scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: No such resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        default:
          description: An error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/tunnels/{tunnelID}:
    get:
      tags:
        - Tunnels
      summary: Get a tunnel
      operationId: getTunnelsByTunnelID
      security:
        - ApiKeyAuth:
            - tunnels:read
      parameters:
        - name: tunnelID
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tunnel'
        "401":
          description: The API key is missing, malformed, revoked or expired.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: The API key lacks the tunnels:read scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: No such resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        default:
          description: An error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Tunnels
      summary: Close a tunnel
      operationId: deleteTunnelsByTunnelID
      security:
        - ApiKeyAuth:
            - tunnels:write
      parameters:
        - name: tunnelID
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Success, with no response body.
        "401":
          description: The API key is missing, malformed, revoked or expired.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: The API key lacks the tunnels:write scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: No such resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        default:
          description: An error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /tunnels/{tunnelID}/:
    get:
      tags:
        - Tunnels
      summary: Tunnel entry point
      description: 'Serves the tunnelled application. Authenticated by the tunnel token (Authorization: Bearer, the X-Marionette-Tunnel-Token header, or HTTP basic with the token as the password) unless the tunnel is public.'
      operationId: openTunnel
      parameters:
        - name: tunnelID
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Whatever the tunnelled service returned.
          content:
            '*/*': {}
        "404":
          description: No such resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        default:
          description: An error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /tunnels/{tunnelID}/{path}:
    get:
      tags:
        - Tunnels
      summary: Tunnel passthrough
      description: Proxies any path, and any method, to the tunnelled service. WebSocket upgrades are relayed too.
      operationId: proxyThroughTunnel
      parameters:
        - name: tunnelID
          in: path
          required: true
          schema:
            type: string
        - name: path
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Whatever the tunnelled service returned.
          content:
            '*/*': {}
        "404":
          description: No such resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        default:
          description: An error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    ApiKeyAuth:
      bearerFormat: mk_...
      description: An API key minted by the admin API.
      scheme: bearer
      type: http
  schemas:
...
    CreateTunnelRequest:
      type: object
      properties:
        session_id:
          type: string
        type:
          type: string
        local_port:
          type: integer
          format: int32
        public:
          type: boolean
      required:
        - type
        - local_port
        - public
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
...
    Tunnel:
      type: object
      properties:
        id:
          type: string
        session_id:
          type: string
        runner_id:
          type: string
        type:
          type: string
          enum:
            - http
            - tcp
            - desktop
            - browser
            - ios
            - android
        direction:
          type: string
          enum:
            - inbound
            - outbound
        local_port:
          type: integer
          format: int32
        public_url:
          type: string
        is_public:
          type: boolean
        token:
          type: string
        expires_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        closed_at:
          type: string
          format: date-time
      required:
        - id
        - session_id
        - type
        - direction
        - local_port
        - is_public
        - expires_at
        - created_at
    TunnelList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Tunnel'
        total_count:
          type: integer
          format: int64
        has_more:
          type: boolean
        next_cursor:
          type: string
      required:
        - items
        - total_count
        - has_more
...